+++ engine/src/utils/types.rs
@@ -6,15 +6,15 @@ use std::rc::Rc;
 use std::sync::{Arc, Mutex, RwLock};
 use std::time::Duration;
 
-pub type Vec2F = cgmath::Vector2<f64>;
+pub type Vec2F = cgmath::Vector2<f32>;
 pub type Vec2I = cgmath::Vector2<i32>;
-pub type Vec3F = cgmath::Vector3<f64>;
-pub type Vec4F = cgmath::Vector4<f64>;
-pub type Mat4F = cgmath::Matrix4<f64>;
-pub type Mat3F = cgmath::Matrix3<f64>;
-pub type QuatF = cgmath::Quaternion<f64>;
-pub type DegF = cgmath::Deg<f64>;
-pub type Mat2F = cgmath::Matrix2<f64>;
+pub type Vec3F = cgmath::Vector3<f32>;
+pub type Vec4F = cgmath::Vector4<f32>;
+pub type Mat4F = cgmath::Matrix4<f32>;
+pub type Mat3F = cgmath::Matrix3<f32>;
+pub type QuatF = cgmath::Quaternion<f32>;
+pub type DegF = cgmath::Deg<f32>;
+pub type Mat2F = cgmath::Matrix2<f32>;
 pub type Color = Vec3F;
 
 pub type Ref<T> = Rc<T>;
